Move ui interfaces to deticated files.
