+++ src/utils/interfaces/ui/ILoaderContext.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from "react";
+
+export interface ILoaderContext {
+    isLoading: boolean;
+    setIsLoading: Dispatch<SetStateAction<boolean>>;
+}
