+++ src/utils/interfaces/ui/IOption.ts
@@ -1,0 +1,4 @@
+export interface IOption {
+    id: string;
+    name: string;
+}
