+++ src/utils/interfaces/ui/ISearchBarProps.ts
@@ -1,0 +1,7 @@
+import { ChangeEventHandler } from "react";
+
+export interface ISearchBarProps {
+    onInputChange: ChangeEventHandler<HTMLInputElement>;
+    defaultValue?: string;
+    placeholder?: string;
+}
