+++ src/utils/interfaces/ui/ISearchFilter.ts
@@ -1,0 +1,6 @@
+export interface ISearchFilter {
+    displayedFilters: { title: string; id: string }[];
+    onFilterClick: (id: string) => void;
+    activeFilter: string;
+    title?: string;
+}
