+++ src/utils/interfaces/ui/ISearchWithFiltersProps.ts
@@ -1,0 +1,8 @@
+import { ISearchBarProps, ISearchFilter, IShouldShowFilters } from ".";
+
+export interface ISearchWithFiltersProps extends ISearchBarProps {
+    subjectFilter: ISearchFilter;
+    topicFilter: ISearchFilter;
+    isLoadingQuestions: boolean;
+    shouldShowFilters: IShouldShowFilters;
+}
