+++ src/utils/interfaces/ui/IShouldShowFilters.ts
@@ -1,0 +1,4 @@
+export interface IShouldShowFilters {
+    subject: boolean;
+    topic: boolean;
+}
