+++ src/utils/interfaces/ui/ISuggestion.ts
@@ -1,0 +1,4 @@
+export interface ISuggestion {
+    name: string;
+    id: string;
+}
