+++ src/utils/interfaces/ui/ITab.ts
@@ -1,0 +1,9 @@
+import { CSSProperties, ReactElement } from "react";
+
+export interface ITab {
+    title: string;
+    content: ReactElement;
+    contentContainerStyle?: CSSProperties;
+    btnStyle?: CSSProperties;
+    tabBtnClickSideEffect?: () => void;
+}
