+++ src/utils/interfaces/ui/UserInteractionFilterType.ts
@@ -1,0 +1,1 @@
+export type UserInteractionFilter = "created" | "answered" | "commented";
